requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
pymongo==4.6.1
schedule==1.2.1
pandas==2.2.0
gspread==5.12.4
oauth2client==4.1.3
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0 